services/novelty_agent/requirements.txt
# NATS client for subscribing to events
nats-py==2.3.0

# OpenAI client for LLM integration
openai==1.12.0
